+++ descwl_shear_sims/psfs/__init__.py
@@ -3,4 +3,4 @@ from . import ps_psf
 from .ps_psf import PowerSpectrumPSF, make_ps_psf
 from .fixed_psf import FixedPSF, make_fixed_psf
 from .rand_psf import make_rand_psf
-from .dmpsfs import FixedDMPSF, PowerSpectrumDMPSF, make_dm_psf
+from .dmpsfs import FixedDMPSF, DMPSF, PowerSpectrumDMPSF, make_dm_psf